Pulls `DynamicMachine` out of `MachineForTarget` and adds `MultiConfigurationTarget` as a first multiplexer.
