+++ Machines/Utility/MachineForTarget.hpp
@@ -11,32 +11,13 @@
 
 #include "../../Analyser/Static/StaticAnalyser.hpp"
 
-#include "../../Configurable/Configurable.hpp"
-#include "../ConfigurationTarget.hpp"
-#include "../CRTMachine.hpp"
-#include "../JoystickMachine.hpp"
-#include "../KeyboardMachine.hpp"
-#include "Typer.hpp"
+#include "../DynamicMachine.hpp"
 
 #include <map>
 #include <string>
 #include <vector>
 
 namespace Machine {
-
-/*!
-	Provides the structure for owning a machine and dynamically casting it as desired without knowledge of
-	the machine's parent class or, therefore, the need to establish a common one.
-*/
-struct DynamicMachine {
-	virtual ~DynamicMachine() {}
-	virtual ConfigurationTarget::Machine *configuration_target() = 0;
-	virtual CRTMachine::Machine *crt_machine() = 0;
-	virtual JoystickMachine::Machine *joystick_machine() = 0;
-	virtual KeyboardMachine::Machine *keyboard_machine() = 0;
-	virtual Configurable::Device *configurable_device() = 0;
-	virtual Utility::TypeRecipient *type_recipient() = 0;
-};
 
 enum class Error {
 	None,
